refactor in-app MetaMask logo replacement with Chainlink logo
