+++ app/component-library/components/Avatars/Avatar/variants/AvatarFavicon/AvatarFavicon.constants.ts
@@ -27,7 +27,7 @@ export const SAMPLE_AVATARFAVICON_SVGIMAGESOURCE_REMOTE: ImageSourcePropType = {
   uri: 'https://metamask.github.io/test-dapp/metamask-fox.svg',
 };
 /* eslint-disable-next-line */
-export const SAMPLE_AVATARFAVICON_IMAGESOURCE_LOCAL: ImageSourcePropType = require('../../../../../../images/fox.png');
+export const SAMPLE_AVATARFAVICON_IMAGESOURCE_LOCAL: ImageSourcePropType = require('../../../../../../images/hexagon.png');
 
 export const SAMPLE_AVATARFAVICON_PROPS: AvatarFaviconProps = {
   imageSource: SAMPLE_AVATARFAVICON_IMAGESOURCE_REMOTE,
