+++ app/components/Nav/Main/MainNavigator.js
@@ -725,7 +725,7 @@ const MainNavigator = () => (
       headerTitle={() => (
         <Image
           style={styles.headerLogo}
-          source={require('../../../images/metamask-name.png')}
+          source={require('../../../images/chainlink-name.png')}
           resizeMode={'contain'}
         />
       )}
