+++ app/components/UI/FoxScreen/index.js
@@ -22,7 +22,7 @@ const createStyles = (colors) =>
     },
   });
 
-const foxImage = require('../../../images/fox.png'); // eslint-disable-line import/no-commonjs
+const foxImage = require('../../../images/hexagon.png'); // eslint-disable-line import/no-commonjs
 
 /**
  * View component that displays the MetaMask fox
