+++ app/components/UI/Notification/Row/Icon.tsx
@@ -13,7 +13,7 @@ import { IconName } from '../../../../component-library/components/Icons/Icon';
 import RemoteImage from '../../../../components/Base/RemoteImage';
 import { ETHEREUM_LOGO } from '../../../../constants/urls';
 /* eslint-disable import/no-commonjs */
-const metamask_fox = require('../../../../images/fox.png'); // eslint-disable-line
+const metamask_fox = require('../../../../images/hexagon.png'); // eslint-disable-line
 interface NotificationIconProps {
   notificationType: TRIGGER_TYPES;
   styles: any;
