+++ app/components/UI/ReviewModal/index.tsx
@@ -36,7 +36,7 @@ const helpOptions: HelpOption[] = [
 ];
 
 /* eslint-disable-next-line */
-const foxImage = require('images/fox.png');
+const foxImage = require('images/hexagon.png');
 
 const ReviewModal = () => {
   const navigation = useNavigation();
