+++ app/components/Views/ChoosePassword/index.js
@@ -574,7 +574,7 @@ class ChoosePassword extends PureComponent {
             <View style={styles.foxWrapper}>
               {Device.isAndroid() ? (
                 <Image
-                  source={require('../../../images/fox.png')}
+                  source={require('../../../images/hexagon.png')}
                   style={styles.image}
                   resizeMethod={'auto'}
                 />
