+++ app/components/Views/Login/index.js
@@ -508,7 +508,7 @@ class Login extends PureComponent {
               <View style={styles.foxWrapper}>
                 {Device.isAndroid() ? (
                   <Image
-                    source={require('../../../images/fox.png')}
+                    source={require('../../../images/hexagon.png')}
                     style={styles.image}
                     resizeMethod={'auto'}
                   />
