+++ app/components/Views/ResetPassword/index.js
@@ -609,7 +609,7 @@ class ResetPassword extends PureComponent {
             <View style={styles.foxWrapper}>
               {Device.isAndroid() ? (
                 <Image
-                  source={require('../../../images/fox.png')}
+                  source={require('../../../images/hexagon.png')}
                   style={styles.image}
                   resizeMethod={'auto'}
                 />
