+++ app/components/Views/Settings/AppInformation/index.js
@@ -85,7 +85,7 @@ const createStyles = (colors) =>
     },
   });
 
-const foxImage = require('../../../../images/fox.png'); // eslint-disable-line import/no-commonjs
+const foxImage = require('../../../../images/hexagon.png'); // eslint-disable-line import/no-commonjs
 
 /**
  * View that contains app information
